% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand.R
\name{blockrand}
\alias{blockrand}
\title{Block randomization}
\usage{
blockrand(seed, rand.block, n)
}
\arguments{
\item{seed}{integer for the random seed generator state.}

\item{rand.block}{values within each block. The length of this vector determine the block size.}

\item{n}{number of observations.}
}
\value{
Vector or binary otcome.
}
\description{
Block randomization
}
\details{
Nothing
}
\examples{
blockrand(123,c(1,1,0,0),10)
}
\author{
Paul Blanche
}
