% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ck.R
\name{calc_ck}
\alias{calc_ck}
\title{Calculate Critval c Using Method 1}
\usage{
calc_ck(
  uk,
  lk,
  Info.i,
  Info.d,
  Info.max = NULL,
  ImaxAnticipated = FALSE,
  rho_alpha = 2,
  alpha = 0.025,
  bindingFutility = TRUE
)
}
\arguments{
\item{uk}{upper bounds for all analyses up to and including current stage k}

\item{lk}{lower bounds for all analyses up to and including current stage k}

\item{Info.i}{Information for all analyses up to and including current stage k}

\item{Info.d}{Observed information at decision analysis k}

\item{Info.max}{Maximum information}

\item{ImaxAnticipated}{set to TRUE if c should be calculated according to Eq 15 in HJ because study was stopped early due to Imax reached}

\item{rho_alpha}{value of rho used for type I error spending function}

\item{alpha}{one-sided alpha level to be used for the study}

\item{bindingFutility}{whether the futility stopping rule is binding}
}
\description{
Calculate Critval c Using Method 1

calculate critval c based on the boundaries at interim.
}
