% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.delayedGSD.R
\name{confint.delayedGSD}
\alias{confint.delayedGSD}
\title{Treatment Effect Estimates for a Group Sequential Design with Delayed Endpoints}
\usage{
\method{confint}{delayedGSD}(
  object,
  parm = NULL,
  level = NULL,
  method = NULL,
  k = NULL,
  type.k = NULL,
  ...
)
}
\arguments{
\item{object}{object of class \code{delayedGSD}, typically output from \code{\link{CalcBoundaries}}.}

\item{parm}{not used, for compatibility with the generic method.}

\item{level}{not used, for compatibility with the generic method.}

\item{method}{[character] type of estimate to output: can be \code{"ML"} or  \code{"corrected ML"}, the latter accounting for the group sequential design.}

\item{k}{[integer] stage relative to which the estimates should be output.}

\item{type.k}{[character] type  of stage relative to which the estimates should be output: \code{"interim"}, \code{"decision"}, or \code{"final"}.}

\item{...}{not used, for compatibility with the generic method.}
}
\description{
Extract estimate, confidence intervals, p-value, ... relative to the treatment effect at a specific stage of a group sequential design with delayed endpoints.
By default extract value for the latest stage that has been performed.
}
