% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{simLP3trial}
\alias{simLP3trial}
\title{Simulated Data Based on a Phase 3 Trial}
\usage{
data(simLP3trial)
}
\description{
Simulated dataset with properties (e.g. correlation structure between repeated measurements and missingness patterns)
similar to a Phase III trial (double-blind, randomized, placebo-controlled, parallel arm) sponsored by H. Lundbeck A/S.
The trial evaluate the efficacy of a drug to treat a certain neurodegenerative disease. Patients are randomized 1:1 to an active and a placebo arm.
The primary outcome is the change from baseline to Week 12 on a continuous endpoint.
This endpoint is measured at Baseline, Week 6 and Week 12.
The endpoint of interest is a score between 0 and 20, where higher scores indicate a worse disease status.

\itemize{
\item \code{id}: study participant.
\item \code{Z}: randomization group (0: placebo arm, 1: active arm)
\item \code{X1}: endpoint measured at baseline (higher is worse).
\item \code{X2}: endpoint measured at week 6 (higher is worse).
\item \code{X3}: endpoint measured at week 12 (higher is worse).
\item \code{t1}: time (in days) elapsed between the start of the trial and the baseline measurement, i.e., when was the patient recruited w.r.t. the start of the study.
\item \code{t2}: time (in days) elapsed between the start of the trial and the week 6 measurement.
\item \code{t3}: time (in days) elapsed between the start of the trial and the week 12 measurement.
}
}
\keyword{datasets}
