% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decision.R
\name{Decision}
\alias{Decision}
\title{Evaluate decision at an interim, decision or final analysis of a group sequential design with delayed endpoints}
\usage{
Decision(
  analysis_res,
  planned_bnds,
  k = 1,
  analysis = "interim",
  Info.i,
  InfoR.d = NULL,
  PositiveIsGood = TRUE,
  Trace = TRUE,
  bindingFutility = TRUE,
  plot = T
)
}
\arguments{
\item{analysis_res}{object of class ?? from AnalyzeData}

\item{planned_bnds}{object of class ?? from CalcBoundaries}

\item{k}{the stage at which the decision is to be made}

\item{analysis}{is it an interim, decision or final analysis}

\item{Info.i}{the observed (where possible) or expected information at each interim and the final analysis}

\item{InfoR.d}{the expected or observed information ratio at each decision analysis}

\item{PositiveIsGood}{whether a positive effect is considered beneficial (TRUE/FALSE)}

\item{Trace}{whether to print some messages}

\item{bindingFutility}{whether to use binding futility boundaries (use TRUE for binding)}

\item{plot}{whether the updated boundaries and the result should be plotted}
}
\value{
ff
}
\description{
Maps the statistical results at an interim, decision or final analysis into a decision regarding whether to stop recruitment (interim) or whether to reject the null hypothesis (decision/final). Stopping boundaries are updated based on observed information and correct p-values, confidence intervals and point estimates are given.
}
\details{
zz
}
\examples{
x <- rnorm(10)
}
\author{
Paul Blanche
}
