% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorSpend.R
\name{ErrorSpend}
\alias{ErrorSpend}
\title{rho-family spending functions (Kim-DeMets) for alpha and beta}
\usage{
ErrorSpend(I, rho, beta_or_alpha, Info.max)
}
\arguments{
\item{I}{Information at analysis to be evaluated}

\item{rho}{parameter for the spending function}

\item{beta_or_alpha}{total alpha or beta to be spent}

\item{Info.max}{Maximum information needed at end of trial}
}
\description{
rho-family spending functions (Kim-DeMets) for alpha and beta
}
