% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalCI.R
\name{FinalCI}
\alias{FinalCI}
\title{calculate confidence intervals at the end of the study}
\usage{
FinalCI(
  Info.d,
  Info.i,
  ck,
  lk,
  uk,
  kMax,
  alpha = 0.05,
  estimate,
  optimizer = "optimise",
  futility2efficacy,
  bindingFutility
)
}
\arguments{
\item{Info.d}{Information at all decision analyses up to stage where study was stopped}

\item{Info.i}{Information at all interim analyses up to stage where study was stopped}

\item{ck}{decision boundaries for all decision analyses up to stage where study was stopped (should include final boundary if stopped at final analysis)}

\item{lk}{lower bounds up to stage where study was stopped}

\item{uk}{upper bounds up to stage where study was stopped}

\item{kMax}{maximum number of analyses}

\item{alpha}{confidence level (to get a 100*(1-alpha)\% CI)}

\item{estimate}{naive estimate (e.g. using  ML or REML).}

\item{optimizer}{the observed treatment estimate at decision}

\item{futility2efficacy}{[logical] is it possible to stop for futility at interim and reject the null hypothesis at decision.}

\item{bindingFutility}{[logical]  whether the futility stopping rule is binding.}
}
\description{
calculate confidence intervals at the end of the study
}
