% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatAsCaseStudy.R
\name{FormatAsCase}
\alias{FormatAsCase}
\title{Convert to L-format}
\usage{
FormatAsCase(d)
}
\arguments{
\item{[data.frame]}{data output by \code{GenData}.}
}
\description{
Rename variable and relevel factor variables.
}
