% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInformation.R
\name{getInformation}
\alias{getInformation}
\title{Extract information relative to a parameter.}
\usage{
getInformation(object, ...)
}
\arguments{
\item{object}{a \code{ttest} object or a \code{gls} object.}

\item{...}{not used. For compatibility with the generic details.}

\item{name.coef}{[character] For which coefficient the information should be computed (only relevant for gls models).}

\item{type}{[character] Should the information be estimated only for the observed data (\code{"estimation"}, excluding missing values)
or for the full dataset  (\code{"prediction"}, including missing values).}

\item{method.prediction}{[character] method used to estimate the information relative to the missing values (only relevant for gls models).
Can be by increasing the information proportionally to the amount of missing values (\code{"inflation"})
or by approximating the variance of the estimator by a weighted average between the complete case estimator and the full information estimator (\code{"pooling"}).
or a list containing the theoretical value of the variance-covariance matrix for each treatment group.}

\item{variance}{[NULL, list] If \code{NULL} then the estimated variance-covariance matrix from the object is used.
Otherwise it should be a list containing the theoretical value of the variance-covariance matrix for each treatment group.}

\item{data}{[data.frame] The dataset relative to which the information should be computed. See details section.}

\item{details}{[logical] Should intermediate results be output. See details section.}
}
\description{
Extract information relative to a parameter.
}
\details{
Argument \bold{data}: the dataset may contain missing value in the outcome but no in the covariates. Missing value in the outcome indicates that the information is not available at the interim analysis but will be come available at decision.


Argument \bold{details}: when using gls models, an attribute detail is added to the output which contain a list:\itemize{
\item decision: information at decision analysis
\item interim: information at the interim analysis using all available data
\item interim.cc: information at the interim analysis using a complete case analysis
\item n: sample size at decision, interim with complete observation, interim with only partial observations
}
}
