% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method1.R
\name{Method1}
\alias{Method1}
\title{calculate critval c using method 1}
\usage{
Method1(
  uk,
  lk,
  Info.i,
  Info.d,
  Info.max = NULL,
  sided = 1,
  cMin = -Inf,
  ImaxAnticipated = FALSE,
  rho = 2,
  alpha = 0.025,
  bindingFutility = TRUE
)
}
\arguments{
\item{uk}{upper bounds for all analyses up to and including current stage k}

\item{lk}{lower bounds for all analyses up to and including current stage k}

\item{Info.i}{Information for all analyses up to and including current stage k}

\item{Info.d}{Observed information at decision analysis k}

\item{Info.max}{Maximum information}

\item{sided}{one or two sided}

\item{cMin}{minimun possible value c for the decision analysis, typically that for a fixed sample test (H & J page 10)}

\item{ImaxAnticipated}{set to TRUE if c should be calculated according to Eq 15 in HJ because study was stopped early due to Imax reached}

\item{rho}{value of rho used for type I error spending function}

\item{alpha}{one-sided alpha level to be used for the study}

\item{bindingFutility}{whether the futility stopping rule is binding}
}
\description{
calculate critval c using method 1
}
