% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method2.R
\name{Method2}
\alias{Method2}
\title{calculate critval c using method 2}
\usage{
Method2(
  uk,
  lk,
  Info.i,
  Info.d,
  Info.max,
  delta,
  sided = 1,
  cMin = -Inf,
  bindingFutility = TRUE
)
}
\arguments{
\item{uk}{upper bounds for all analyses up to and including current stage k}

\item{lk}{lower bounds for all analyses up to and including current stage k}

\item{Info.i}{Information for all analyses up to and including current stage k}

\item{Info.d}{Observed information at decision analysis k}

\item{Info.max}{Maximum information}

\item{delta}{planned effect}

\item{sided}{one or two sided}

\item{cMin}{minimun possible value c for the decision analysis, typically that for a fixed sample test (H & J page 10)}

\item{bindingFutility}{whether the futility stopping rule is binding}
}
\description{
calculate critval c using method 2
}
