% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonBindingHJ.R
\name{NonBindingHJ}
\alias{NonBindingHJ}
\title{Calculate boundaries for a group sequential design}
\usage{
NonBindingHJ(
  rho_alpha = 2,
  rho_beta = 2,
  alpha = 0.025,
  beta = 0.2,
  Kmax,
  Info.max = NULL,
  Info.i = NULL,
  Info.d = NULL,
  theta = 0,
  abseps = 1e-06,
  binding = TRUE,
  binding_type = "BBO",
  direction = "smaller",
  Trace = FALSE,
  nWhileMax = 30,
  toldiff = 1e-05,
  tolcoef = 1e-04,
  mycoefMax = 1.2,
  mycoefL = 1,
  myseed = 2902
)
}
\arguments{
\item{rho_alpha}{rho parameter of the rho-family spending functions (Kim-DeMets) for alpha}

\item{rho_beta}{rho parameter of the rho-family spending functions (Kim-DeMets) for beta}

\item{alpha}{type I error}

\item{beta}{type II error}

\item{Kmax}{max number of analyses (including final)}

\item{Info.max}{maximum information needed for given beta (type II-error), theta (expected difference), alpha (type I-error) and Kmax. It can be given if it is known. Otherwise it is computed from the  values given for alpha, beta, theta and Kmax.}

\item{Info.i}{Expected or observed (wherever possible) information at the interim and final analyses 1:Kmax}

\item{Info.d}{Expected or observed information at all potential decision analyses 1:(Kmax-1)}

\item{theta}{expected effect under the alternative (should be on the scale of the test statistc for which If and Info.max relate to one over the variance, e.g. theta=expected log(Hazard ratio))}

\item{abseps}{tolerance for precision when finding roots or computing integrals}

\item{binding}{binding or non-binding futility boundary (i.e FALSE if it is allowed to continue after crossing the futility boundary)}

\item{binding_type}{type of non-binding futility rule to use. "HJ" corresponds to the method proposed by Hampson and Jennison, "BBO" corresponds to the method proposed by Baayen, Blanche and Ozenne}

\item{direction}{greater is for Ho= theta > 0, "smaller" is for Ho= theta < 0 (note that in Jennison and Turnbull's book chapter (2013) they consider smaller)}

\item{Trace}{Used only if Info.max=NULL. Whether to print informations to follow the progression of the (root finding) algorithm to compute Info.max (from  alpha, beta, theta and Kmax).}

\item{nWhileMax}{Used only if Info.max=NULL. Maximum number of steps in the (root finding) algorithm to compute Info.max (from  alpha, beta, theta and Kmax)}

\item{toldiff}{Used only if Info.max=NULL. Maximum tolerated difference between lower and upper bounds at anaylis Kmax (which souhld be zero), in the root finding algorithm, to find the value of Info.max}

\item{tolcoef}{Used only if Info.max=NULL. Maximum tolerated difference before stopping the search (in the root finding algorithm), between two successive values for the multiplier coeficient 'coef' such that Info.max=coef*If  (some values for coef are given in Table 7.6 page 164 Jennison's book. The value of "If" (which stands for Information for Fixed design) corresponds to the information we need if Kmax=1)}

\item{mycoefMax}{Used only if Info.max=NULL. Upper limit of the interval of values in which we search for the multiplier coeficient 'coef' such that Info.max=coef*If (in the root finding algorithm).}

\item{mycoefL}{Used only if Info.max=NULL. Lower limit of the interval (see mycoefMax)}

\item{myseed}{seed for producing reproducible results. Because we call functions which are based on Monte-Carlo compuation (pmvnorm)}
}
\description{
Calculate boundaries for a group sequential design without delayed endpoints based on planned and/or observed information using an error spending approach. The function can also give the boundaries for a non-binding futility rule for delayed endpoints following a method proposed by Hampson and Jennison.
}
\examples{

## Example to check that code matches
b1 <- CalcBoundaries(kMax=2,  #max number of analyses (including final)
                    sided=1,  #one or two-sided
                    alpha=0.025,  #type I error
                    beta=0.2,  #type II error
                    InfoR.i=c(0.6,1),  #planned information rates
                    gammaA=2,  #rho parameter for alpha error spending function
                    gammaB=2,  #rho parameter for beta error spending function
                    method=1,  #use method 1 or 2 from paper H&J
                    delta=1.5,  #effect that the study is powered for
                    InfoR.d=0.65,
                    bindingFutility=FALSE)

NonBindingHJ(Kmax=2,Info.max=b1$Info.max,theta=1.5,binding=FALSE,alpha=0.025,Info.i=c(0.6,1)*b1$Info.max)
}
