% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingDelayedGSD.R
\name{operatingDelayedGSD}
\alias{operatingDelayedGSD}
\title{Operating Characteristics of GSD via simulation}
\usage{
operatingDelayedGSD(
  n.obs = NULL,
  n.sim,
  method,
  kMax,
  InfoR.i,
  InfoR.d,
  delta,
  args.GenData = NULL,
  alpha = 0.025,
  beta = 0.1,
  rho_alpha = 2,
  rho_beta = 2,
  PropForInterim = InfoR.i,
  lag = 0,
  path = NULL,
  name = "simGSD",
  export.tempo = TRUE,
  trace = TRUE,
  seed = NULL,
  cpus = 1
)
}
\arguments{
\item{n.obs}{[integer vector, >0] total sample size in the trial.
If \code{NULL}, the sample size corresponds to the type II error of the first method will be used.}

\item{n.sim}{[integer vector, >0] number of simulations.
Can also be of length 2, in that case the second element indicates which replicate of the simulation is being run.}

\item{method}{[data.frame] method(s) to be used.
Should be a data.frame containing 3 columns: method (integer from 1 to 3), binding (logical), fixC (logical).
An additional column \code{overrule.futility} can be added to indicate what to do in presence of non-binding futility rules (default: never stop for futility under the null, always stop for futility under the alternative).}

\item{kMax}{[integer] max number of analyses (including final).}

\item{InfoR.i}{[numeric vector of length kMax] planned information rates at interim and final.}

\item{InfoR.d}{[numeric vector of length kMax] planned information rates at decision and final.}

\item{delta}{[numeric] planned treatment effect (i.e. difference in mean) on primary outcome.}

\item{args.GenData}{[list] arguments passed to the function GenData used to generate datasets.}

\item{alpha}{[numeric, 0-1] type I error (one sided).}

\item{beta}{[numeric, 0-1] type II error (one sided).}

\item{rho_alpha}{[numeric, >0] rho parameter for alpha error spending function.}

\item{rho_beta}{[numeric, >0] rho parameter for beta error spending function.}

\item{PropForInterim}{[numeric vector of length Kmax-1] percentage of all subjects, once they had the change to have complete follow-up, triggering interim and final analyses.
By default equal to \code{InfoR.i}.}

\item{lag}{[numeric] time lag between stop of recruitment and decision to stop recruitment.}

\item{path}{[character] directory where to export the results.
If \code{NULL}, will be exported in the current directory.
If \code{NA}, will not be exported.}

\item{name}{[character] start of the filename.}

\item{export.tempo}{[logical] should results be exported at each iteration in a temporary file.}

\item{trace}{[logical] When \code{TRUE}, ouput is generate in the console as the execution of the function progresses.}

\item{seed}{[integer, >0] Random number generator (RNG) state used when starting resampling. The same seed is used for all methods.}

\item{cpus}{[integer, >0] number of child-processes for parallel evaluation.}
}
\description{
Simulate trials with delayed endpoint and output the operating characteristics of a GSD.
}
\examples{

df.method <- data.frame(method = 1:3, binding = TRUE, fixC = c(FALSE,FALSE,TRUE))

MyMissProb <- matrix(c(0.04807692, 0.05769231, 0.00961538, 0.88461538),  nrow = 2, ncol = 2,
                     dimnames = list(c("V1 missing", "V1 not missing"),c("V2 missing", "V2 not missing")))

args.GenData <- list(rand.block = c(1, 1, 0, 0),
                     allsd = c(2.5, 2.1, 2.4),
                     mean0 = c(10, 0, 0),
                     delta = c(0, 0.5, 1)*1,
                     ar = 30,
                     cor.01.1 = -0.15,
                     cor.ij.1 = 0.68,
                     cor.0j.1 = -0.27,
                     MissProb = MyMissProb,
                     DigitsOutcome = 2,
                     TimeFactor = 42,
                     DigitsTime = 0)

##### 2 stages #### 
res2stage <- operatingDelayedGSD(n.sim = 10, 
             method = df.method,
             args.GenData = args.GenData,
             kMax = 2, InfoR.i = c(0.56,1), InfoR.d = c(0.65, 1), delta = 1,
             PropForInterim = 0.5, lag = 21,
             seed = 1:10)

##### 3 stages ####
res3stage <- operatingDelayedGSD(n.sim = 10, 
             method = df.method[1,,drop=FALSE],
             args.GenData = args.GenData,
             kMax = 3, InfoR.i = c(0.40,0.65,1), InfoR.d = c(0.50,0.75, 1), delta = 1,
             PropForInterim = c(0.35,0.6), lag = 21,
             seed = 1:10)

}
