% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.delayedGSD.R
\name{plot.delayedGSD}
\alias{plot.delayedGSD}
\title{Display boundaries over information}
\usage{
\method{plot}{delayedGSD}(
  x,
  y,
  type = "Z",
  Itype = "rate",
  planned = NULL,
  predicted = TRUE,
  main = NULL,
  xlim = NULL,
  ylim = NULL,
  legend = TRUE,
  legend.x = NULL,
  legend.ncol = 1,
  legend.cex = 1,
  add.arrow = FALSE,
  sep.arrow = c(0.2, 0.1),
  size.arrow = 0.1,
  lwd.arrow = 1,
  cex.bound = 1.2,
  cex.estimate = 1.2,
  ...
)
}
\arguments{
\item{x}{object from CalcBoundaries}

\item{y}{not used. For compatibility with the generic method.}

\item{type}{type of boundaries to plot (Z-statistic (Z), p-value (P), effect (E))}

\item{Itype}{information scale on x-axis (rate or absolute (abs))}

\item{planned}{[logical] should the planned boundaries be displayed along with the boundaries that have been computed based on the data.
Can also be \code{"only"} to only display planned boundaries.}

\item{predicted}{[logical] Should the predicted information/boundaries at decision based on interim data be output (when relevant).}

\item{main}{to specify a plot title}

\item{xlim}{range of the x axis}

\item{ylim}{range of the y axis}

\item{legend}{should a caption be added?}

\item{legend.x}{location of the legend on the plot.}

\item{legend.ncol}{number of column in the caption}

\item{legend.cex}{character expansion factor for the text in the legend.}

\item{add.arrow}{[logical] should arrows indicating the ordering of the space be added?}

\item{sep.arrow}{[numeric vector of length 2] space between the critical point and the arrow (below and above)}

\item{size.arrow}{[numeric, >0] width of the arrow.}

\item{lwd.arrow}{[numeric, >0] thickness of the arrow.}

\item{cex.bound}{[numeric, >0] size of the dots for the boundaries.}

\item{cex.estimate}{[numeric vector of size 2] size of the dots for the estimated statistic/p-value/effect and its label.}

\item{...}{not used. For compatibility with the generic method.}
}
\description{
Display boundaries over information
}
\examples{
#### planning stage ####
bnds3 <- CalcBoundaries(kMax=3, InfoR.i = c(0.5,0.75,1), InfoR.d = c(0.55,0.8))
plot(bnds3, type = "Z")
plot(bnds3, type = "Z", add.arrow = TRUE)
plot(bnds3, type = "Z", add.arrow = TRUE, lwd.arrow=2,ylim=c(0.35,2.6), xlim=c(0.5,1.1),legend=FALSE)
plot(bnds3, type = "E")
plot(bnds3, type = "P")

GSD.bound <- CalcBoundaries(kMax=2,  ## max number of analyses (including final)
                              alpha=alpha,  ## type I error
                              beta=beta,  ## type II error
                              InfoR.i=c(round(I_i,2),1),  ## planned or observed information rates
                              rho_alpha=2,  ## rho parameter for alpha error spending function
                              rho_beta=2,  ## rho parameter for beta error spending function
                              method=3,  ## use method 1 or 2 from paper H&J
                              delta=theta0,  ## effect that the study is powered for
                              InfoR.d=c(round(I_d,2),1),
                              cNotBelowFixedc=TRUE)

#### analysis stage ####
data(simLP3trial, package = "DelayedGSD")
resI <- analyzeData(SelectData(simLP3trial,453.6779))
GSD.analysis <- update(GSD.bound,delta=resI)
plot(GSD.analysis, type = "Z")
plot(GSD.analysis, type = "E")
plot(GSD.analysis, type = "P")
}
