% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gridDelayedGSD.R
\name{plot.gridDelayedGSD}
\alias{plot.gridDelayedGSD}
\title{Display p-values}
\usage{
\method{plot}{gridDelayedGSD}(
  object,
  col = NULL,
  xlab = "Stage",
  ylab = "Test statistic",
  title = NULL,
  xlim = NULL,
  pch = 20,
  cex = 2,
  cex.lab = 1.2,
  cex.p = 1,
  font.p = 2,
  lwd = 2,
  digits = 2
)
}
\arguments{
\item{object}{outpout of gridDelayedGSD.}

\item{col}{[character vector] color used to display the arrows.}

\item{xlab}{[character] label for the x-axis.}

\item{ylab}{[character] label for the y-axis.}

\item{title}{[character] title of the plot.}

\item{xlim}{[numeric vector of length 2] minimum and maximum value on the x-axis.}

\item{pch}{[integer, >0] shape used to display the boundaries at decision and final.}

\item{cex}{[numeric, >0] size of the shape used to display the boundaries at decision and final.}

\item{cex.lab}{[numeric, >0] size of the x-axis and y-axis labels.}

\item{cex.p}{[numeric, >0] size of the used to display the p-values.}

\item{font.p}{[numeric, >0] font used to display the p-values.}

\item{digits}{[integer, >0] number of digit to be used when displaying the p-values.}

\item{ylim}{[numeric vector of length 2] minimum and maximum value on the y-axis.}
}
\description{
Display p-value over the sample space.
}
