% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProgress.R
\name{PlotProgress}
\alias{PlotProgress}
\title{Display Sample Size over Time}
\usage{
PlotProgress(d, Delta.t = 0.500001, at = NA)
}
\arguments{
\item{d}{dataset}

\item{Delta.t}{??}

\item{at}{??}
}
\description{
Display the available and pipeline sample size as a function of time
}
