% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.delayedGSD.R
\name{print.delayedGSD}
\alias{print.delayedGSD}
\title{Status of a Group Sequential Design with Delayed Endpoints}
\usage{
\method{print}{delayedGSD}(x, ...)
}
\arguments{
\item{x}{object of class \code{delayedGSD}, typically output from \code{\link{CalcBoundaries}}.}

\item{...}{arguments passed to the summary function.}
}
\description{
Display boundaries and estimated treatment effect up the current stage.
}
