% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.operatingDelayedGSD.R
\name{print.operatingDelayedGSD}
\alias{print.operatingDelayedGSD}
\title{Simulation settings used to assess GSD Operating Characteristics}
\usage{
\method{print}{operatingDelayedGSD}(x, index.method = 1, print = TRUE, digits = c(2, 2), ...)
}
\arguments{
\item{x}{output of operatingDelayedGSD.}

\item{index.method}{[integer,>=1] index of the method relative to which the information and occurence of a stage should be display.}

\item{print}{[logical] should the information be printed in the console.}

\item{digits}{[integer vector of length 2] number of digits used to round, respectively, percentage and numbers.}

\item{...}{not used, for compatibility with the generic method}
}
\description{
Display key simulation parameters (sample size, occurence of a stage, information) of the simulation study used to assess the operating characteristics of a GSD.
}
