% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.delayedGSD.R
\name{summary.delayedGSD}
\alias{summary.delayedGSD}
\title{Status of a Group Sequential Design with Delayed Endpoints}
\usage{
\method{summary}{delayedGSD}(
  x,
  planned = NULL,
  predicted = TRUE,
  digits = 5,
  space = "|",
  abreviated = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class \code{delayedGSD}, typically output from \code{\link{CalcBoundaries}}.}

\item{planned}{[logical] should the planned boundaries be displayed along with the boundaries that have been computed based on the data.
Can also be \code{"only"} to only display planned boundaries.}

\item{predicted}{[logical] Should the predicted information/boundaries at decision based on interim data be output (when relevant).}

\item{digits}{[interger,>0] integer indicating the number of decimal places to be used.}

\item{space}{[character] column separator.}

\item{abreviated}{[logical] abreviate some column names and action that should be taken
(e.g. C for continue, C-Info for continue due to decreasing information, S-F for stop for futility, and S-Imax for stop because maximum information has been reached).}

\item{...}{not used, for compatibility with the generic method.}
}
\description{
Display boundaries and estimated treatment effect up the current stage.
}
