% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.operatingDelayedGSD.R
\name{summary.operatingDelayedGSD}
\alias{summary.operatingDelayedGSD}
\title{Summary of the Operating Characteristics of a GSD}
\usage{
\method{summary}{operatingDelayedGSD}(object, print = TRUE, digits = c(2, 4), ...)
}
\arguments{
\item{print}{[logical] should the information be printed in the console.}

\item{digits}{[logical] should the information be printed in the console.}

\item{...}{not used, for compatibility with the generic method}

\item{x}{output of operatingDelayedGSD.}
}
\description{
Display key simulation parameters (sample size, occurence of a stage, information) and the key operating characteristics of a GSD.
}
