% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest}
\alias{ttest}
\title{Student's t-Test}
\usage{
ttest(...)
}
\arguments{
\item{...}{arguments passed to \code{stats::t.test}}
}
\description{
Performs one and two sample t-tests on vectors of data. 
The only modification to the \code{stats::t.test} is that it stores the call and the arguments.
}
