% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.delayedGSD.R
\name{update.delayedGSD}
\alias{update.delayedGSD}
\title{Perform Next Analysis Stage in a GSD}
\usage{
\method{update}{delayedGSD}(
  object,
  data,
  PositiveIsGood = NULL,
  ddf = NULL,
  k = NULL,
  type.k = NULL,
  trace = TRUE,
  ...
)
}
\arguments{
\item{object}{object of class \code{delayedGSD}, typically output from \code{\link{CalcBoundaries}}.}

\item{data}{[data.frame] dataset used to fit the linear mixed model.}

\item{PositiveIsGood}{[logical] whether a positive effect is considered beneficial.}

\item{ddf}{[character] method used to compute the degrees of freedom of the test statistics. Argument passed to \code{\link{analyzeData}}.}

\item{k}{[integer] index of the analysis.}

\item{type.k}{[character] type of analysis: \code{"interim"} (after continuing recruitment),
\code{"decision"} (after stopping recruitment for efficacy or futility),
or \code{"final"} (after reaching the last stage of the trial).}

\item{trace}{[logical] should the execution of the function be traced?}
}
\description{
Use the newly observed data to perform the next stage of the analysis:
fit a mixed model, update the information, recompute the boundaries, take a decision, and possibly correct the estimated treatment effect.
}
\examples{

#### Planning #####
theAlpha <- 0.025
theBeta <- 0.2
theDelta <- 1.5
theK <- 2
theN <- 82

myBound0 <- CalcBoundaries(kMax=theK,
                          sided=1,
                     alpha=theAlpha,
                     beta=theBeta,
                     InfoR.i=c(0.5,1),
                     gammaA=2,
                     gammaB=2,
                     method=1,
                     delta=theDelta,
                     InfoR.d=0.55)

#### Simulate data ####
set.seed(10)
theData <- GenData(n=theN*2,delta=theDelta*0.8,ar=5)  #generate data with all data for in case trial completes

theAR <- 10  #accrual rate (pt per month)
theDelay <- 0.7500001  #time in months to process data
tau.i <- theData$d$t3[theN + ceiling(theAR*theDelay)] #time point at which to do IA


#### Analyse data at the first interim ####
theInterimData <- SelectData(theData$d, t = tau.i, Delta.t = theDelay)

myInterim1 <- update(myBound0, data = theInterimData, k = 1, analysis = "interim")
print(myInterim1)
print(myInterim1, planned = FALSE)
print(myInterim1, planned = "only")

par(mfrow = c(1,2))
plot(myInterim1, planned = "only")
plot(myInterim1)

#### Analyse data at the final stage ####
theFinalData <- SelectData(theData$d, t = 1e7, Delta.t = theDelay) 
myFinal <- update(myInterim1, data = theFinalData, k = 2, analysis = "final")
}
