% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInformation.R
\name{updateInformation}
\alias{updateInformation}
\title{Update Information in a GSD}
\usage{
updateInformation(object, lmm, k, type.k, update.stage = TRUE)
}
\arguments{
\item{lmm}{Linear mixed model as a \code{lmmGSD} object, typically output from \code{\link{analyzeData}}.}

\item{k}{[integer] Index of the analysis.}

\item{type.k}{[character] Type of analysis: \code{"interim"} (after continuing recruitment),
\code{"decision"} (after stopping recruitment for efficacy or futility),
or \code{"final"} (after reaching the last stage of the trial).}

\item{update.stage}{[logical] should the arguments \code{k} and \code{type.k} be used to update to stage of the trial?}

\item{x}{Object of type \code{delayedGSD}, typically output from \code{\link{CalcBoundaries}}.}
}
\description{
Update the information based according to a linear mixed model.
}
