% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeData.R
\name{analyzeData}
\alias{analyzeData}
\title{Fit a Linear Mixed Model}
\usage{
analyzeData(
  data,
  ddf = "nlme",
  getinfo = TRUE,
  data.decision = NULL,
  trace = TRUE
)
}
\arguments{
\item{data}{[data.frame] dataset for the current analysis.}

\item{ddf}{[character] method used to compute the degrees of freedom of the Wald statistic.
Can be \code{"satterthwaite"} or \code{"nlme"}.}

\item{getinfo}{[logical] should the information be computed at interim and decision? Otherwise no information is computed.}

\item{data.decision}{[integer or data.frame] data or total number of patients for the future decision analysis (only relevant at interim).
Used to compute the predicted total information at decision where only current (or future) drop-out value are excluded but future observations are kept.}

\item{trace}{[logical] should the execution of the function be traced?}
}
\description{
Fit a linear mixed model and extract the information relative to the parameter of interest.
}
\examples{
## simulate data in the wide format
set.seed(1)
dataW <- GenData(n=104)$d

## estimate mixed model
analyzeData(dataW, getinfo = FALSE)
analyzeData(dataW, getinfo = TRUE)
analyzeData(dataW, ddf = "satterthwaite", getinfo = TRUE)

}
