% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand.R
\name{blockrand}
\alias{blockrand}
\title{Block randomization}
\usage{
blockrand(seed, rand.block, n)
}
\value{
Vector or binary otcome.
}
\description{
Block randomization
}
\details{
Nothing
}
\examples{
blockrand(123,c(1,1,0,0),10)
}
\author{
Paul Blanche
}
