% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcBoundaries.R
\name{CalcBoundaries}
\alias{CalcBoundaries}
\title{Calculate boundaries for a group sequential design with delayed endpoints}
\usage{
CalcBoundaries(
  kMax = 2,
  alpha = 0.025,
  beta = 0.2,
  InfoR.i = c(0.5, 1),
  rho_alpha = 2,
  rho_beta = 2,
  method = 1,
  cNotBelowFixedc = FALSE,
  delta = 1.5,
  InfoR.d = 0.55,
  bindingFutility = TRUE,
  alternative = "greater",
  conf.level = 1 - 2 * alpha,
  n = NA,
  trace = FALSE,
  mycoefMax = 1.2,
  PowerCorrection = FALSE
)
}
\arguments{
\item{kMax}{max number of analyses (including final)}

\item{alpha}{type I error}

\item{beta}{type II error}

\item{InfoR.i}{planned or observed information rates at interim analysis, including the final analysis.}

\item{rho_alpha}{rho parameter for alpha error spending function}

\item{rho_beta}{rho parameter for beta error spending function}

\item{method}{use method 1 or 2 from paper H&J}

\item{cNotBelowFixedc}{whether the value c at the decision analysis can be below that of a fixed sample test (H & J page 10)}

\item{delta}{effect that the study is powered for}

\item{InfoR.d}{(expected) information rate at each decision analysis (i.e. when stopping at an interim analysis). Should not include the final analysis.}

\item{bindingFutility}{[logical]  whether the futility stopping rule is binding.}

\item{alternative}{a character string specifying the alternative hypothesis, \code{"greater"} or \code{"less"}.
H0 \eqn{\theta=0} vs H1 \eqn{theta<0} (\code{"less"}) or theta > 0 (\code{"greater"}).
Note that in Jennison and Turnbull's book chapter (2013) they consider greater.}

\item{conf.level}{confidence level for the confidence intervals. By default the complement to 1 of twice the type 1 error.}

\item{n}{planned sample size in each group. Optional argument.}

\item{trace}{whether to print some messages}

\item{PowerCorrection}{whether to correct the power for Method 1}
}
\description{
Calculate boundaries (interim and decision) for a group sequential design with delayed endpoints based on planned and/or observed information using an error spending approach
}
\examples{
myBound <- CalcBoundaries(kMax=2,
              alpha=0.025,  
              beta=0.2,  
              InfoR.i=c(0.5,1),
              rho_alpha=2,
              rho_beta=2,
              method=2,
              cNotBelowFixedc=TRUE,
              delta=1.5,
              InfoR.d=0.55)
myBound
plot(myBound)

## to reproduce bounds from CJ DSBS course slide 106
myBound <- CalcBoundaries(kMax=3,
              alpha=0.025,  
              beta=0.1,  
              InfoR.i=c(3.5,6.75,12)/12,
              rho_alpha=1.345,
              rho_beta=1.345,
              method=1, ## has been changed from 2 to 1
              cNotBelowFixedc=TRUE,
              bindingFutility=FALSE,
              delta=1,
              InfoR.d=c(5.5,8.75)/12)
myBound
plot(myBound)
}
