% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decision.R
\name{Decision}
\alias{Decision}
\title{Take Decision in a GSD}
\usage{
Decision(object, k, type.k, PositiveIsGood = TRUE, trace = TRUE)
}
\arguments{
\item{object}{object of class \code{delayedGSD}, typically output from \code{\link{CalcBoundaries}}.}

\item{k}{[integer] index of the analysis.}

\item{type.k}{[character] type of analysis: \code{"interim"} (after continuing recruitment),
\code{"decision"} (after stopping recruitment for efficacy or futility),
or \code{"final"} (after reaching the last stage of the trial).}

\item{PositiveIsGood}{[logical] whether a positive effect is considered beneficial (TRUE/FALSE)}

\item{trace}{[logical] should the execution of the function be traced?}
}
\value{
ff
}
\description{
Maps the statistical results at an interim, decision or final analysis into a decision regarding whether to stop recruitment (interim) or whether to reject the null hypothesis (decision/final). Stopping boundaries are updated based on observed information and correct p-values, confidence intervals and point estimates are given.
}
\details{
Function that maps the statistical results at interim (or final) analysis into decision to reject, or continue or stop inclusind subjects.
}
\examples{

#### Planning #####
theAlpha <- 0.025
theBeta <- 0.2
theDelta <- 1.5
theK <- 2
theN <- 82

myBound0 <- CalcBoundaries(kMax=theK,
                     sided=1,
                     alpha=theAlpha,
                     beta=theBeta,
                     InfoR.i=c(0.5,1),
                     gammaA=2,
                     gammaB=2,
                     method=1,
                     delta=theDelta,
                     InfoR.d=0.55)

#### Simulate data ####
## generate data with all data for in case trial completes
set.seed(10)
theData <- GenData(n=theN*2,delta=theDelta*0.8,ar=5)  

theAR <- 10  #accrual rate (pt per month)
theDelay <- 0.7500001  #time in months to process data
tau.i <- theData$d$t3[theN + ceiling(theAR*theDelay)] #time point at which to do IA

theObsData <- SelectData(theData$d, t = tau.i, Delta.t = theDelay)

#### Analyse data at interim ####
myLMM <- analyzeData(theObsData)
myBound1 <- updateBoundaries(myBound0, lmm = myLMM, k = 1, type.k = "interim")
myInterim1 <- Decision(myBound1, k = 1, type.k = "interim")
myInterim1$conclusion
}
\author{
Paul Blanche
}
