% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedGSD.options.R
\name{DelayedGSD.options}
\alias{DelayedGSD.options}
\title{Global options for DelayedGSD package}
\usage{
DelayedGSD.options(..., reinitialise = FALSE)
}
\arguments{
\item{...}{options to be selected or updated}

\item{reinitialise}{should all the global parameters be set to their default value}
}
\value{
A list containing the default options.
}
\description{
Update or select global options for the DelayedGSD package.
}
\details{
The options are: \itemize{
\item FCT.p_value [character]: function to evaluate p-values (FinalPvalue or FinalPvalue2)
\item continuity.correction [0,1,2]: correction used to evaluate the p-value when using \code{cNotBelowFixedc=TRUE}. Positive values ensures continuity of the p-value across stages.
\item tolerance [numeric,>0]: acceptable discrepancy to the objective level when evaluating the confidence intervals and median unbiased estimate.
}
}
\examples{
DelayedGSD.options()
}
\keyword{utilities}
