% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalCI.R
\name{FinalCI}
\alias{FinalCI}
\title{calculate confidence intervals at the end of the study}
\usage{
FinalCI(
  Info.d,
  Info.i,
  ck,
  lk,
  uk,
  sided = 1,
  kMax,
  alpha = 0.05,
  estimate,
  optimizer = "optimise"
)
}
\arguments{
\item{Info.d}{Information at all decision analyses up to stage where study was stopped}

\item{Info.i}{Information at all interim analyses up to stage where study was stopped}

\item{ck}{decision boundaries for all decision analyses up to stage where study was stopped (should include final boundary if stopped at final analysis)}

\item{lk}{lower bounds up to stage where study was stopped}

\item{uk}{upper bounds up to stage where study was stopped}

\item{sided}{whether the test is 1 or 2-sided}

\item{kMax}{maximum number of analyses}

\item{alpha}{confidence level (to get a 100*(1-alpha)\% CI)}

\item{estimate}{naive estimate (e.g. using  ML or REML).}

\item{optimizer}{the observed treatment estimate at decision}
}
\description{
calculate confidence intervals at the end of the study
}
