% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalEstimate.R
\name{FinalEstimate}
\alias{FinalEstimate}
\title{calculate de-biased estimate}
\usage{
FinalEstimate(
  Info.d,
  Info.i,
  ck,
  ck.unrestricted,
  lk,
  uk,
  kMax,
  estimate,
  method,
  bindingFutility,
  cNotBelowFixedc
)
}
\arguments{
\item{Info.d}{Information at all decision analyses up to stage where study was stopped (should include information at final analysis if stopped at final analysis)}

\item{Info.i}{Information at all interim analyses up to stage where study was stopped}

\item{ck, ck.unrestricted}{decision boundaries for all decision analyses up to stage where study was stopped (should include final boundary if stopped at final analysis).
ck is possibly with restriction (when cNotBelowFixedc=TRUE) and ck.unrestricted always without.}

\item{lk}{lower bounds up to stage where study was stopped}

\item{uk}{upper bounds up to stage where study was stopped}

\item{kMax}{maximum number of analyses}

\item{estimate}{naive estimate (e.g. using  ML or REML).}

\item{method}{method 1, 2 or 3}

\item{bindingFutility}{[logical]  whether the futility stopping rule is binding.}

\item{cNotBelowFixedc}{[logical] whether the value c at the decision analysis can be below that of a fixed sample test (H & J page 10)}
}
\description{
calculate de-biased estimate
}
