% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.delayedGSD.R
\name{plot.delayedGSD}
\alias{plot.delayedGSD}
\title{Display boundaries over information}
\usage{
\method{plot}{delayedGSD}(
  x,
  y,
  type = "Z",
  Itype = "rate",
  planned = TRUE,
  main = NULL,
  xlim = NULL,
  ylim = NULL,
  legend = TRUE,
  legend.ncol = 1,
  legend.cex = 1,
  add.arrows = FALSE,
  sep.arrow = c(0.2, 0.1),
  size.arrow = 0.1,
  lwd.arrow = 1,
  ...
)
}
\arguments{
\item{x}{object from CalcBoundaries}

\item{y}{not used. For compatibility with the generic method.}

\item{type}{type of boundaries to plot (Z-statistic (Z), p-value (P), effect (E))}

\item{Itype}{information scale on x-axis (rate or absolute (abs))}

\item{main}{to specify a plot title}

\item{xlim}{range of the x axis}

\item{ylim}{range of the y axis}

\item{legend}{should a caption be added?}

\item{legend.ncol}{number of column in the caption}

\item{legend.cex}{character expansion factor for the text in the legend.}

\item{add.arrows}{[logical] should arrows indicating the ordering of the space be added?}

\item{...}{not used. For compatibility with the generic method.}

\item{sep.arrows}{[numeric vector of length 2] space between the critical point and the arrow (below and above)}

\item{size.arrows}{[numeric, >0] width of the arrow.}

\item{lwd.arrows}{[numeric, >0] thickness of the arrow.}
}
\description{
Display boundaries over information
}
\examples{
bnds <- CalcBoundaries(kMax=3, InfoR.i = c(0.5,0.75,1), InfoR.d = c(0.55,0.8))
plot(bnds, type = "Z")
plot(bnds, type = "Z", add.arrows = TRUE)
plot(bnds, type = "Z", add.arrows = TRUE, lwd.arrow=2,ylim=c(0.35,2.6), xlim=c(0.5,1.1),legend=FALSE)
plot(bnds, type = "E")
plot(bnds, type = "P")
}
